refactor: move usernamestatus to seprate file
